api/requirements.txt
Flask==2.2.2
inject==4.3.1
bcrypt==4.0.1
pyjwt==2.6.0
python-dotenv==0.21.0
dataclasses-json==0.5.7
